Fixed NPE in LSMBTreeUpdateTest.
git-svn-id: https://hyracks.googlecode.com/svn/branches/hyracks_lsm_tree@1184 123451ca-8445-de46-9d55-352943316053

+++ hyracks-tests/hyracks-storage-am-lsm-btree-test/src/test/java/edu/uci/ics/hyracks/storage/am/lsm/btree/util/LSMBTreeTestHarness.java
@@ -45,7 +45,6 @@ public class LSMBTreeTestHarness {
     private static final long RANDOM_SEED = 50;
     private static final int DEFAULT_DISK_PAGE_SIZE = 256;
     private static final int DEFAULT_DISK_NUM_PAGES = 1000;
-    //private static final int DEFAULT_DISK_NUM_PAGES = 100;
     private static final int DEFAULT_DISK_MAX_OPEN_FILES = 200;
     private static final int DEFAULT_MEM_PAGE_SIZE = 256;
     private static final int DEFAULT_MEM_NUM_PAGES = 100;
@@ -113,9 +112,11 @@ public class LSMBTreeTestHarness {
                 }
             };
             String[] files = dir.list(filter);
-            for (String fileName : files) {
-                File file = new File(dir.getPath() + File.separator + fileName);
-                file.delete();
+            if (files != null) {
+                for (String fileName : files) {
+                    File file = new File(dir.getPath() + File.separator + fileName);
+                    file.delete();
+                }
             }
             dir.delete();
         }
